pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoscaling-forecasting"
version = "0.1.0"
description = "Time series forecasting for autoscaling decisions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "lightgbm>=4.0",
    "xgboost>=2.0",
    "prophet>=1.1",
    "tensorflow>=2.13",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "streamlit>=1.28",
    "plotly>=5.17",
    "altair>=5.0"
]
[tool.setuptools]
packages = ["forecasting"]